pyproject.toml
[project]
name = "homelab-dns-exporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "prometheus-client>=0.22.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
]
